MarzneshinIpLimit/requirements.txt
websockets==13.1
python-telegram-bot==21.6
Flask==3.0.3
python-jose==3.3.0